# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Chat API
  version: "1"
  description: |
    success: {code:0,message:"ok",data:<schema>}
    error: {code:http_status,message,error_code}
servers:
  - { url: 'http://127.0.0.1:8081' }
  - { url: 'http://127.0.0.1:8082' }
  - { url: 'http://127.0.0.1:8083' }
  - { url: 'http://127.0.0.1:28081' }
  - { url: 'http://127.0.0.1:28082' }
  - { url: 'http://127.0.0.1:28083' }
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Err:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
        error_code: { type: string }
    Status:
      type: object
      properties:
        status: { type: string }
    U:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string }
    F:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string }
        remark_name: { type: string }
        is_ai_bot: { type: boolean }
...
    FR:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string }
        direction: { type: string }
        message: { type: string }
        requester: { $ref: '#/components/schemas/U' }
        addressee: { $ref: '#/components/schemas/U' }
        created_at: { type: string }
        updated_at: { type: string }
    C:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string }
        name: { type: string }
        announcement: { type: string }
        creator_id: { type: integer, format: int64 }
        member_count: { type: integer }
        pinned: { type: boolean }
        pinned_at: { type: string }
        is_muted: { type: boolean }
        draft: { type: string }
        last_read_seq: { type: integer, format: int64 }
        unread_count: { type: integer }
        last_message_seq: { type: integer, format: int64 }
        last_message_sender: { type: integer, format: int64 }
        last_message_preview: { type: string }
        last_message_type: { type: string }
        last_message_at: { type: string }
    CM:
      type: object
      properties:
        user_id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        avatar_url: { type: string }
        remark_name: { type: string }
        role: { type: string }
        last_read_seq: { type: integer, format: int64 }
        joined_at: { type: string }
        online: { type: boolean }
...
    UP:
      type: object
      required: [display_name]
      properties:
        display_name: { type: string }
        avatar_url: { type: string }
    CFR:
      type: object
      required: [addressee_id]
      properties:
        addressee_id: { type: integer, format: int64 }
        message: { type: string }
    UFR:
      type: object
      properties:
        remark_name: { type: string }
    CDC:
      type: object
      required: [target_user_id]
      properties:
        target_user_id: { type: integer, format: int64 }
    CGC:
      type: object
      required: [name, member_ids]
      properties:
        name: { type: string }
        member_ids:
          type: array
          items: { type: integer, format: int64 }
    RC:
      type: object
      required: [name]
      properties:
        name: { type: string }
    ACM:
      type: object
      required: [member_ids]
      properties:
        member_ids:
          type: array
          items: { type: integer, format: int64 }
...
    MR:
      type: object
      properties:
        seq: { type: integer, format: int64 }
...
paths:
...
  /api/v1/users/me:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=U
          content: { application/json: { schema: { $ref: '#/components/schemas/U' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
    patch:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UP' } } }
      responses:
        '200':
          description: data=U
          content: { application/json: { schema: { $ref: '#/components/schemas/U' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/users:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=U[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/U' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/friends:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=F[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/F' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/friends/{id}/remark:
    patch:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UFR' } } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/friend-requests:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=FR[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FR' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CFR' } } }
      responses:
        '200':
          description: data=FR
          content: { application/json: { schema: { $ref: '#/components/schemas/FR' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/friend-requests/{id}/approve:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/friend-requests/{id}/reject:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=C[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/C' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/direct:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CDC' } } }
      responses:
        '200':
          description: data=C
          content: { application/json: { schema: { $ref: '#/components/schemas/C' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/group:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CGC' } } }
      responses:
        '200':
          description: data=C
          content: { application/json: { schema: { $ref: '#/components/schemas/C' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}:
    patch:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RC' } } }
      responses:
        '200':
          description: data=C
          content: { application/json: { schema: { $ref: '#/components/schemas/C' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/members:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=CM[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CM' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ACM' } } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/members/{userId}:
    delete:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: path
          name: userId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/leave:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/transfer-owner:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TO' } } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/settings:
    patch:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UCS' } } }
      responses:
        '200':
          description: data=C
          content: { application/json: { schema: { $ref: '#/components/schemas/C' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/messages:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: cursor
          required: false
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer }
      responses:
        '200':
          description: data=M[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/M' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MSG' } } }
      responses:
        '200':
          description: data=M
          content: { application/json: { schema: { $ref: '#/components/schemas/M' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/conversations/{id}/read:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/MR' } } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
...
  /api/v1/messages/{id}/recall:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/uploads/presign:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UPR' } } }
      responses:
        '200':
          description: data=UPS
          content: { application/json: { schema: { $ref: '#/components/schemas/UPS' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/uploads/object/{key}:
    put:
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '204':
          description: uploaded
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/uploads/complete:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UCR' } } }
      responses:
        '200':
          description: data=UCSP
          content: { application/json: { schema: { $ref: '#/components/schemas/UCSP' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/search:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: q, required: false, schema: { type: string } }
        - { in: query, name: scope, required: false, schema: { type: string, enum: [all, messages, conversations] } }
        - { in: query, name: cursor, required: false, schema: { type: string } }
        - { in: query, name: limit, required: false, schema: { type: integer } }
        - { in: query, name: conversation_id, required: false, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=SR
          content: { application/json: { schema: { $ref: '#/components/schemas/SR' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/sync/cursor:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=SC
          content: { application/json: { schema: { $ref: '#/components/schemas/SC' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/sync/events:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: cursor, required: false, schema: { type: integer, format: int64 } }
        - { in: query, name: limit, required: false, schema: { type: integer } }
      responses:
        '200':
          description: data=SER
          content: { application/json: { schema: { $ref: '#/components/schemas/SER' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/auth/login:
    post:
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminLogin' } } }
      responses:
        '200':
          description: data=AdminAuth
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminAuth' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/auth/me:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=AdminProfile
          content: { application/json: { schema: { $ref: '#/components/schemas/AdminProfile' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/auth/change-password:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AdminPwd' } } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai-config:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=AIConfig
          content: { application/json: { schema: { $ref: '#/components/schemas/AIConfig' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
    put:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AIConfig' } } }
      responses:
        '200':
          description: data=AIConfig
          content: { application/json: { schema: { $ref: '#/components/schemas/AIConfig' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai/retry-jobs:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: limit, required: false, schema: { type: integer } }
        - { in: query, name: status, required: false, schema: { type: string } }
      responses:
        '200':
          description: data=AIRetryJob[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AIRetryJob' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai/retry-jobs/{id}:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=AIRetryJob
          content: { application/json: { schema: { $ref: '#/components/schemas/AIRetryJob' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai/retry-jobs/{id}/retry-now:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai/retry-jobs/retry-batch:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items: { type: integer, format: int64 }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/ai/retry-jobs/cleanup:
    post:
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                statuses:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/monitor/overview:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=MonitorOverview
          content: { application/json: { schema: { $ref: '#/components/schemas/MonitorOverview' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/monitor/timeseries:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=MonitorTimeseries
          content: { application/json: { schema: { $ref: '#/components/schemas/MonitorTimeseries' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/audit/ai-calls:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: limit, required: false, schema: { type: integer } }
        - { in: query, name: status, required: false, schema: { type: string } }
        - { in: query, name: provider, required: false, schema: { type: string } }
        - { in: query, name: model, required: false, schema: { type: string } }
        - { in: query, name: user_id, required: false, schema: { type: integer, format: int64 } }
        - { in: query, name: conversation_id, required: false, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=AIAudit[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AIAudit' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/audit/ai-calls/{id}:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=AIAuditDetail
          content: { application/json: { schema: { $ref: '#/components/schemas/AIAuditDetail' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/messages/resolve:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: query, name: client_msg_id, required: true, schema: { type: string } }
        - { in: query, name: sender_id, required: false, schema: { type: integer, format: int64 } }
        - { in: query, name: conversation_id, required: false, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=MessageLookup
          content: { application/json: { schema: { $ref: '#/components/schemas/MessageLookup' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/message-journey/{messageId}:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: messageId, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=MessageJourney
          content: { application/json: { schema: { $ref: '#/components/schemas/MessageJourney' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/conversations/{id}/consistency:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: data=Consistency
          content: { application/json: { schema: { $ref: '#/components/schemas/Consistency' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/conversations/{id}/events:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer, format: int64 } }
        - { in: query, name: limit, required: false, schema: { type: integer } }
      responses:
        '200':
          description: data=SE[]
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SE' }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
  /api/v1/admin/search/reindex:
    post:
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: data=Status
          content: { application/json: { schema: { $ref: '#/components/schemas/Status' } } }
        default:
          description: error
          content: { application/json: { schema: { $ref: '#/components/schemas/Err' } } }
...